+++ examples/data-sources/ceph_role/data-source.tf
@@ -1,0 +1,3 @@
+data "ceph_role" "administrator" {
+  name = "administrator"
+}