+++ examples/data-sources/ceph_user/data-source.tf
@@ -1,0 +1,3 @@
+data "ceph_user" "admin" {
+  username = "admin"
+}