+++ internal/provider/roles_data_source_test.go
@@ -12,20 +12,18 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 )
 
-func TestAccMonitorsDataSource(t *testing.T) {
-	const addr = "data.ceph_monitors.all"
+func TestAccRolesDataSource(t *testing.T) {
+	const addr = "data.ceph_roles.all"
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			{
-				Config: `data "ceph_monitors" "all" {}`,
+				Config: `data "ceph_roles" "all" {}`,
 				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttrWith(addr, "monitors.#", atLeastOne),
-					resource.TestCheckResourceAttrSet(addr, "monitors.0.name"),
-					resource.TestCheckResourceAttrSet(addr, "monitors.0.public_addr"),
-					resource.TestCheckResourceAttrSet(addr, "monitors.0.in_quorum"),
+					resource.TestCheckResourceAttrWith(addr, "roles.#", atLeastOne),
+					resource.TestCheckResourceAttrSet(addr, "roles.0.name"),
 				),
 			},
 		},